Staff Homepage App Route
Right now, this is basically the reusage of the given "/home code", but with the home_staff.html. Only done so that it can run using the pre-existing code and the page can be visible for editing. When changing the redirect(url_for()) function at line 165 from "home" to "home_staff", the page should display without error.