File for Staff Homepage
The staff has a menu bar with different, clickable options, a nice welcome message, and a useful tool to be used when dealing with the scheduling of flights. The menu bar is not operational at the moment, more for proof of concept at the time of uploading. "Home" will take the user back to the home_staff page, "Aerial System" will include mechanisms to create big changes to the system like adding a new airport,  "Payroll" is self-explanatory since it's about money, and "About" is honestly just a bullshit page for aesthetics but it shouldn't be deleted. Further down, "Flight Navigator" displays information from the database based on the option selected from the dropdown menu. I plan on taking out anything that modifies the data from the drop-down menu and placing them as buttons margined to the right, below the data. Logout works without issues.